requirements.txt
attrs==21.4.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
click
codecovopentelem==0.0.6
coverage==6.4.1
Deprecated==1.2.13
Flask==1.1.4
gevent==21.8.0
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.4
iniconfig==1.1.1
itsdangerous
Jinja2==2.11.3
markupsafe==2.0.1
opentelemetry-api==1.11.1
opentelemetry-instrumentation==0.30b1
opentelemetry-instrumentation-flask==0.30b1
opentelemetry-instrumentation-wsgi==0.30b1
opentelemetry-sdk==1.11.1
opentelemetry-semantic-conventions==0.30b1
opentelemetry-util-http==0.30b1
packaging==21.3
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.9
pytest==7.1.2
pytest-cov==3.0.0
requests==2.27.1
tomli==2.0.1
typing_extensions==4.2.0
urllib3==1.26.9
Werkzeug
wrapt==1.14.1
zipp==3.8.0
